+++ src/models/transaction/request/evm.rs
@@ -141,7 +141,7 @@ pub fn validate_price_params(
     if is_legacy {
         if let RelayerNetworkPolicy::Evm(evm_policy) = &relayer.policies {
             if let Some(gas_price_cap) = evm_policy.gas_price_cap {
-                if request.gas_price.unwrap_or(0) > gas_price_cap as u128 {
+                if request.gas_price.unwrap_or(0) > gas_price_cap {
                     return Err(ApiError::BadRequest("Gas price is too high".to_string()));
                 }
             }
